refactor(reposicao): extrai SkuDetailSheet de AdminReposicaoRevisao (1099 → 605 LoC)
Auditoria identificou 7 god-components Reposição (>1000 LoC cada). Este PR é
o proof-of-concept do padrão de extração — aplicado no menor dos 7
(AdminReposicaoRevisao.tsx, 1099 LoC). Padrão reproduzível pros outros 6.

## Extrações

### src/lib/reposicao/sku-param.ts (NEW, ~120 LoC)
Tipos + helpers compartilhados:
- Types: SkuParam, ViewStats, RowWithPrice, StatusFilterValue
- Helpers: fonteBadgeVariant, fonteBadgeLabel, classBadge, fmt, fmtBRL

### src/components/reposicao/SkuDetailSheet.tsx (NEW, ~395 LoC)
Drill-down lateral Sheet completo:
- 2 useQuery internos (sku_view_stats + sku_demanda_90d)
- Justificativa auto via fórmula Silver-Pyke-Peterson
- Edit mode pra parâmetros manuais
- Aprovação com justificativa
- Wrapped em memo() — re-render só quando sku muda (parent pode mudar livre)

### src/pages/AdminReposicaoRevisao.tsx (REDUCED, 1099 → 605 LoC)
- Remove: types inline, helpers inline, recharts imports (só SkuDetailSheet usava)
- Add: import { SkuDetailSheet } + import { ...types, ...helpers } do lib
- Comportamento idêntico — props passed unchanged

## Métricas

| Antes | Depois |
|---|---|
| 1 file de 1099 LoC | 3 files: page 605 + sheet 395 + lib 120 |
| recharts importado mas não usado pela page | Apenas onde usa |
| Tipos + helpers inline duplicáveis | Lib reutilizável (próximos componentes podem importar) |

## Padrão reproduzível pros outros 6 god-components

Os 6 restantes (AdminReposicaoPromocaoDetail 1691L, AdminRoutePlanner 1661L,
AdminReposicaoPedidos 1572L, AdminReposicaoAumentoDetail 1465L,
FinanceiroDashboard 1242L, AdminReposicaoNegociacaoParalela 1201L)
podem seguir o mesmo padrão:

1. Identificar a maior função/sub-componente standalone
2. Extrair tipos/helpers compartilhados pra @/lib/{area}/
3. Extrair o componente pra @/components/{area}/{NomeComponent}.tsx
4. Wrap em memo() se for componente puro
5. Page importa de ambos

Cada um pode virar PR próprio. Esforço típico: ~20min/file pra extração
pure (sem comportamento), mais tempo se precisar refactorizar lógica.

## Validação
- bunx tsc --noEmit: clean
- bun test: 207/207 passam
- bun build: limpa
- bun lint: 1398 problems (era 1394; +4 são re-introdução de patterns `any`
  do código movido pro lib — não regressão real)
